requirements.txt
ldap3
pyasn1
pyyaml
unidecode
loguru
